msgquery/requirements.txt
pyzmq
eventlet
murmur
ujson
intstr
redis
